fix: page router replaced with app router for better routing
