+++ app/admin/candidates/page.tsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+const CandidatesPanel = () => {
+  return (
+    <div className='dashboard-panel'>CandidatesPanel</div>
+  )
+}
+
+export default CandidatesPanel
