+++ app/admin/email/page.tsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+const EmailPanel = () => {
+  return (
+    <div className='dashboard-panel'>EmailPanel</div>
+  )
+}
+
+export default EmailPanel
