+++ app/admin/organizations/page.tsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+const OrganizationPanel = () => {
+  return (
+    <div className='dashboard-panel'>OrganizationPanel</div>
+  )
+}
+
+export default OrganizationPanel
