+++ app/admin/settings/page.tsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+const SettingsPanel = () => {
+  return (
+    <div className='dashboard-panel'>SettingsPanel</div>
+  )
+}
+
+export default SettingsPanel
