+++ app/auth/admin/login/page.tsx
@@ -1,0 +1,8 @@
+import React from "react";
+import LoginForm from "src/components/forms/login";
+
+const AdminLogin = () => {
+  return <LoginForm userRole="admin" />;
+};
+
+export default AdminLogin;
