+++ app/auth/admin/signup/page.tsx
@@ -1,0 +1,10 @@
+import React from "react";
+import SignUpForm from "@/components/forms/signup";
+
+const AdminSignup = () => {
+  return (
+      <SignUpForm userRole="admin" />
+  );
+};
+
+export default AdminSignup;
