+++ app/auth/candidate/signup/page.tsx
@@ -1,0 +1,10 @@
+import React from "react";
+import SignUpForm from "@/components/forms/signup";
+
+const AdminSignup = () => {
+  return (
+      <SignUpForm userRole="candidate" />
+  );
+};
+
+export default AdminSignup;
