--- pages/auth/admin/signup.tsx
@@ -1,13 +1,0 @@
-import React from "react";
-import AuthLayout from "../layout";
-import SignUpForm from "@/components/forms/signup";
-
-const AdminSignup = () => {
-  return (
-    <AuthLayout imgSrc="">
-      <SignUpForm userRole="admin" />
-    </AuthLayout>
-  );
-};
-
-export default AdminSignup;
