+++ src/components/panels/home/emails-table.tsx
@@ -1,3 +1,5 @@
+"use client";
+
 import React, { useState } from "react";
 import Card from "../../ui/card/card";
 import { Switch } from "@headlessui/react";
