+++ src/components/ui/sidebar.tsx
@@ -1,3 +1,5 @@
+"use client";
+
 import React, { useState } from "react";
 import logoTransparent from "@/assets/logo-transparent.png";
 import { GoHomeFill } from "react-icons/go";
@@ -17,40 +19,41 @@ import {
 } from "./dialog/dialog";
 import Button from "./button/button";
 import Link from "next/link";
+import { usePathname, useRouter } from "next/navigation";
 
 const DashboardSidebar: React.FC = () => {
   const clearAuth = useAuthStore((state) => state.clearAuth);
-  // const location = useLocation();
-
+  const router = useRouter();
+  const pathname = usePathname();
   const menuItems = [
     {
       title: "Dashboard",
-      to: "/admin/dashboard/home",
+      to: "/admin/dashboard/",
       icon: <GoHomeFill />,
     },
     {
       title: "Assessments",
-      to: "/admin/dashboard/assessments",
+      to: "/admin/assessments",
       icon: <FaWpforms />,
     },
     {
       title: "Organizations",
-      to: "/admin/dashboard/organizations",
+      to: "/admin/organizations",
       icon: <PiBuildingApartmentFill />,
     },
     {
       title: "Candidates",
-      to: "/admin/dashboard/candidates",
+      to: "/admin/candidates",
       icon: <IoPeople />,
     },
     {
       title: "Email Management",
-      to: "/admin/dashboard/email",
+      to: "/admin/email",
       icon: <PiBuildingApartmentFill />,
     },
     {
       title: "Settings",
-      to: "/admin/dashboard/settings",
+      to: "/admin/settings",
       icon: <PiBuildingApartmentFill />,
     },
   ];
@@ -59,18 +62,18 @@ const DashboardSidebar: React.FC = () => {
 
   return (
     <>
-      <div className="w-1/5 h-full bg-sky-900 hidden lg:block backdrop-blur-lg shadow-2xl z-100">
+      {/* Desktop sidebar */}
+      <aside className="w-1/5 h-screen sticky top-0 bg-sky-900 hidden lg:block backdrop-blur-lg shadow-2xl z-100">
         <div className="flex gap-1 justify-center border-b-2 border-secondary rounded-b-xl items-center">
           <img className="w-20 -ml-3 mt-2" src={logoTransparent.src} />
           <h2 className="text-primary mt-3">Admin Portal</h2>
         </div>
         <div className="w-full flex flex-col px-4 mt-10 gap-5 items-center text-secondary">
           {menuItems.map((item, idx) => {
-            const isActive = true;
-
+            const isActive = pathname.startsWith(item.to);
             return (
               <Link
-                href={item.to ?? ""}
+                href={item.to}
                 key={idx}
                 className={`w-full no-underline p-2 rounded-xl transition duration-300 flex gap-2 items-center justify-start ${
                   isActive
@@ -78,8 +81,8 @@ const DashboardSidebar: React.FC = () => {
                     : "hover:bg-secondary hover:text-sky-900 hover:translate-x-2"
                 }`}
               >
-                <p>{item.icon}</p>
-                <p>{item.title}</p>
+                <span>{item.icon}</span>
+                <span>{item.title}</span>
               </Link>
             );
           })}
@@ -87,10 +90,8 @@ const DashboardSidebar: React.FC = () => {
           <Dialog>
             <DialogTrigger asChild>
               <div className="w-full no-underline hover:bg-secondary shadow-secondary hover:shadow-md hover:text-sky-900 p-2 rounded-xl transition duration-300 hover:translate-x-2 flex gap-2 items-center justify-start cursor-pointer">
-                <p>
-                  <IoLogOut />
-                </p>
-                <p>Logout</p>
+                <IoLogOut />
+                <span>Logout</span>
               </div>
             </DialogTrigger>
             <DialogContent>
@@ -111,9 +112,9 @@ const DashboardSidebar: React.FC = () => {
             </DialogContent>
           </Dialog>
         </div>
-      </div>
+      </aside>
 
-      {/* Mobile collapsible sidebar */}
+      {/* Mobile toggle button */}
       <div
         className={`absolute z-50 -translate-y-1/2 top-1/2 ${
           expandMobileMenu ? "right-1/5" : "left-0"
@@ -126,14 +127,16 @@ const DashboardSidebar: React.FC = () => {
           {!expandMobileMenu ? <FaChevronRight /> : <FaChevronLeft />}
         </button>
       </div>
+
+      {/* Mobile slide-out sidebar */}
       <AnimatePresence>
         {expandMobileMenu && (
-          <motion.div
+          <motion.aside
             initial={{ opacity: 0, x: "-100%" }}
             animate={{ opacity: 1, x: 0 }}
-            transition={{ duration: 0.2, ease: "backInOut" }}
             exit={{ x: "-100%", opacity: 0 }}
-            className="fixed w-3/4 h-full bg-sky-900 lg:hidden backdrop-blur-lg shadow-2xl z-100"
+            transition={{ duration: 0.2, ease: "backInOut" }}
+            className="fixed top-0 left-0 w-3/4 h-screen bg-sky-900 lg:hidden backdrop-blur-lg shadow-2xl z-100"
           >
             <div className="flex gap-1 justify-center border-b-2 border-secondary rounded-b-xl items-center">
               <img className="w-20 -ml-3 mt-2" src={logoTransparent.src} />
@@ -141,21 +144,20 @@ const DashboardSidebar: React.FC = () => {
             </div>
             <div className="w-full flex flex-col px-6 mt-10 gap-5 items-center text-secondary">
               {menuItems.map((item, idx) => {
-                const isActive = true;
-
+                const isActive = pathname.startsWith(item.to);
                 return (
                   <Link
-                    onClick={() => setExpandMobileMenu(false)}
-                    href={item.to!}
+                    href={item.to}
                     key={idx}
+                    onClick={() => setExpandMobileMenu(false)}
                     className={`w-full no-underline p-2 rounded-xl transition duration-300 flex gap-2 items-center justify-start ${
                       isActive
                         ? "bg-secondary text-sky-900 translate-x-2 shadow-secondary shadow-md"
                         : "hover:bg-secondary hover:text-sky-900 hover:translate-x-2"
                     }`}
                   >
-                    <p>{item.icon}</p>
-                    <p>{item.title}</p>
+                    <span>{item.icon}</span>
+                    <span>{item.title}</span>
                   </Link>
                 );
               })}
@@ -163,10 +165,8 @@ const DashboardSidebar: React.FC = () => {
               <Dialog>
                 <DialogTrigger asChild>
                   <div className="w-full no-underline hover:bg-secondary shadow-secondary hover:shadow-md hover:text-sky-900 p-2 rounded-xl transition duration-300 hover:translate-x-2 flex gap-2 items-center justify-start cursor-pointer">
-                    <p>
-                      <IoLogOut />
-                    </p>
-                    <p>Logout</p>
+                    <IoLogOut />
+                    <span>Logout</span>
                   </div>
                 </DialogTrigger>
                 <DialogContent>
@@ -187,7 +187,7 @@ const DashboardSidebar: React.FC = () => {
                 </DialogContent>
               </Dialog>
             </div>
-          </motion.div>
+          </motion.aside>
         )}
       </AnimatePresence>
     </>
