+++ styles/globals.css
@@ -37,7 +37,7 @@ td {
 }
 
 .card {
-    @apply backdrop-blur-sm bg-white/70 shadow-lg rounded-xl;
+    @apply backdrop-blur-sm border-2 border-gray-200 bg-white/70 shadow-lg rounded-xl;
 }
 
 .glass-container {
@@ -45,7 +45,7 @@ td {
 }
 
 .dashboard-panel {
-    @apply relative w-full h-full overflow-auto flex flex-col px-10 py-5 flex-nowrap gap-5 items-start justify-start;
+    @apply relative w-full h-full overflow-auto flex flex-col px-10 py-10 flex-nowrap gap-5 items-start justify-start;
 }
 
 .candidate {
